LangChain/requirements.txt
# Core Flask dependencies
Flask==3.1.2
flask-cors==6.0.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# Environment management
python-dotenv==1.2.1

# LangChain and AI dependencies
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.5
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
openai==1.109.1

# Additional langchain ecosystem (installed as dependencies)
langchain-classic==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9

# HTTP and networking
httpx==0.28.1
httpx-sse==0.4.3
httpcore==1.0.9
requests==2.32.5
requests-toolbelt==1.0.0
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
h11==0.16.0

# Data handling and serialization
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
dataclasses-json==0.6.7
marshmallow==3.26.1
jsonpatch==1.33
jsonpointer==3.0.0
PyYAML==6.0.3
orjson==3.11.4
ormsgpack==1.12.0
zstandard==0.25.0
xxhash==3.6.0

# Async and concurrency
aiohttp==3.13.2
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
async-timeout==4.0.3
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0
anyio==4.11.0
sniffio==1.3.1

# Utilities
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
mypy_extensions==1.1.0
attrs==25.4.0
annotated-types==0.7.0
exceptiongroup==1.3.0
packaging==23.2
regex==2025.11.3
tqdm==4.67.1
tenacity==8.5.0

# OpenAI specific
tiktoken==0.5.2
jiter==0.11.1
distro==1.9.0

# SQL and database
SQLAlchemy==2.0.44

# Scientific computing (if needed for future features)
numpy==1.26.4
